Forecast Task

This task specializes mlr3::Task, mlr3::TaskSupervised and mlr3::TaskRegr for forecasting problems. The target column is assumed to be numeric. The task_type is set to "fcst".

It is recommended to use as_task_fcst() for construction. Predefined tasks are stored in the dictionary mlr3::mlr_tasks.

See also

Super classes

mlr3::Task -> mlr3::TaskSupervised -> mlr3::TaskRegr -> TaskFcst

Public fields

freq

(character(1))
The frequency of the time series.

Active bindings

properties

(character())
Set of task properties. Possible properties are are stored in mlr_reflections$task_properties. The following properties are currently standardized and understood by tasks in mlr3:

"strata": The task is resampled using one or more stratification variables (role "stratum").

"groups": The task comes with grouping/blocking information (role "group").

"weights": The task comes with observation weights (role "weight").

"ordered": The task has columns which define the row order (role "order").

"keys": The task has columns which define the time series "key").

Note that above listed properties are calculated from the $col_roles and may not be set explicitly.

order

(data.table::data.table())
If the task has a column with designated role "order", a table with two or more columns:

row_id (integer()), and

order (Date() | POSIXct() | numeric()).

Returns NULL if there are is no order column.

key

(data.table::data.table())
If the task has a column with designated role "key", a table with two or more columns:

row_id (integer()), and

key variable(s) (factor()).

If there is only one key column, it will be named as key. Returns NULL if there are are no key columns.

Methods

Public methods

TaskFcst$new()

TaskFcst$data()

TaskFcst$print()

TaskFcst$clone()

Inherited methods

mlr3::Task$add_strata()

mlr3::Task$cbind()

mlr3::Task$divide()

mlr3::Task$droplevels()

mlr3::Task$filter()

mlr3::Task$format()

mlr3::Task$formula()

mlr3::Task$head()

mlr3::Task$help()

mlr3::Task$levels()

mlr3::Task$missings()

mlr3::Task$rbind()

mlr3::Task$rename()

mlr3::Task$select()

mlr3::Task$set_col_roles()

mlr3::Task$set_levels()

mlr3::Task$set_row_roles()

mlr3::TaskRegr$truth()

Method new()

Creates a new instance of this R6 class. The function as_task_fcst() provides an alternative way to construct forecast tasks.

Usage

TaskFcst$new(
  id,
  backend,
  target,
  order,
  key = NULL,
  freq = NULL,
  label = NA_character_,
  extra_args = list()
)

Arguments

id

(character(1))
Identifier for the new instance.

backend

(mlr3::DataBackend)
Either a mlr3::DataBackend, or any object which is convertible to a mlr3::DataBackend with as_data_backend(). E.g., a data.frame() will be converted to a mlr3::DataBackendDataTable.

target

(character(1))
Name of the target column.

order

(character(1))
Name of the order column.

key

(character(1))
Name of the key column.

freq

(character(1) | integer(1))
Frequency of the time series. Either a positive number or one of the following:

"secondly"

"minutely"

"hourly"

"daily"

"weekly"

"monthly"

"quarterly"

"yearly"

label

(character(1))
Label for the new instance.

extra_args

(named list())
Named list of constructor arguments, required for converting task types via mlr3::convert_task().

Method data()

Returns a slice of the data from the mlr3::DataBackend as a data.table. Rows default to observations with role "use", and columns default to features with roles "target", "order" or "feature". If rows or cols are specified which do not exist in the mlr3::DataBackend, an exception is raised.

Rows and columns are returned in the order specified via the arguments rows and cols. If rows is NULL, rows are returned in the order of task$row_ids. If cols is NULL, the column order defaults to c(task$target_names, task$feature_names). Note that it is recommended to not rely on the order of columns, and instead always address columns with their respective column name.

Usage

TaskFcst$data(rows = NULL, cols = NULL, ordered = FALSE)

Arguments

rows

(positive integer())
Vector or row indices.

cols

(character())
Vector of column names.

ordered

(logical(1))
If TRUE, data is ordered according to the columns with column role "order".

Returns

Depending on the mlr3::DataBackend, but usually a data.table::data.table().

Method print()

Printer.

Usage

TaskFcst$print(...)

Arguments

...

(ignored).

Method clone()

The objects of this class are cloneable with this method.

Usage

TaskFcst$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(data.table)
airpassengers = tsbox::ts_dt(AirPassengers)
setnames(airpassengers, c("month", "passengers"))
task = as_task_fcst(airpassengers, target = "passengers", order = "month", freq = "monthly")
task$task_type
#> [1] "fcst"
task$formula()
#> passengers ~ .
#> NULL
task$truth()
#> Error in colnamesInt(x, neworder, check_dups = FALSE, skip_absent = skip_absent): argument specifying columns received non-existing column(s): cols[1]='month'