ETS

ETS model. Calls forecast::ets() from package forecast.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("fcst.ets")
lrn("fcst.ets")

Meta Information

Task type: “fcst”

Predict Types: “response”, “quantiles”

Feature Types: “integer”, “numeric”, “Date”

Required Packages: mlr3, mlr3forecast, forecast

Parameters

Id	Type	Default	Levels	Range

model	untyped	"ZZZ"		-

damped	logical	NULL	TRUE, FALSE	-

alpha	numeric	NULL		\((-\infty, \infty)\)

beta	numeric	NULL		\((-\infty, \infty)\)

gamma	numeric	NULL		\((-\infty, \infty)\)

phi	numeric	NULL		\((-\infty, \infty)\)

additive.only	logical	FALSE	TRUE, FALSE	-

lambda	untyped	-		-

biasadj	logical	FALSE	TRUE, FALSE	-

lower	untyped	c(rep(1e-04, 3), 0.8)		-

upper	untyped	c(rep(0.9999, 3), 0.98)		-

opt.crit	character	lik	lik, amse, mse, sigma, mae	-

nmse	integer	3		\([0, 30]\)

bounds	character	both	both, usual, admissible	-

ic	character	aicc	aicc, aic, bic	-

restrict	logical	TRUE	TRUE, FALSE	-

allow.multiplicative.trend	logical	FALSE	TRUE, FALSE	-

na.action	character	na.contiguous	na.contiguous, na.interp, na.fail	-

References

Hyndman, R.J., Koehler, A.B., Snyder, R.D., Grose, S. (2002). “A state space framework for automatic forecasting using exponential smoothing methods.” International J. Forecasting, 18(3), 439–454.

Hyndman, R.J., Akram, Md., Archibald, B. (2008). “The admissible parameter space for exponential smoothing models.” Annals of Statistical Mathematics, 60(2), 407–426.

Hyndman, R.J., Koehler, A.B., Ord, J.K., Snyder, R.D. (2008). Forecasting with exponential smoothing: the state space approach. Springer-Verlag. http://www.exponentialsmoothing.net.

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> mlr3forecast::LearnerFcst -> mlr3forecast::LearnerFcstForecast -> LearnerFcstEts

Methods

Public methods

LearnerFcstEts$new()

LearnerFcstEts$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerFcstEts$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerFcstEts$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.