Time Series Feature Extraction (catch22)

This PipeOp extracts the 22 (or 24) canonical time series characteristics (catch22) from the target variable. For more details, see Rcatch22::catch22_all(), which is called internally on the ordered target vector.

The catch22 set is a low-redundancy subset of the hctsa features, selected for their performance across a diverse collection of time series classification tasks, but applicable as general-purpose features for other tasks such as regression.

For other time series feature extractors, see PipeOpFcstTsfeats and PipeOpFcstFeasts.

Parameters

The parameters are the parameters inherited from mlr3pipelines::PipeOpTaskPreproc, as well as:

catch24 :: logical(1)
If TRUE, additionally compute the mean and standard deviation (the catch24 set). Default FALSE.

Naming

The new columns are named {target}_catch22_{feature}. If the target was called "y" and the feature is "DN_HistogramMode_5", the corresponding new column will be called "y_catch22_DN_HistogramMode_5".

Super classes

mlr3pipelines::PipeOp -> mlr3pipelines::PipeOpTaskPreproc -> PipeOpFcstCatch22

Methods

Public methods

PipeOpFcstCatch22$new()

PipeOpFcstCatch22$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

PipeOpFcstCatch22$new()

Initializes a new instance of this Class.

Usage

PipeOpFcstCatch22$new(id = "fcst.catch22", param_vals = list())

Arguments

id

(character(1))
Identifier of resulting object, default "fcst.catch22".

param_vals

(named list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

PipeOpFcstCatch22$clone()

The objects of this class are cloneable with this method.

Usage

PipeOpFcstCatch22$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(mlr3pipelines)
task = tsk("airpassengers")
po = po("fcst.catch22")
out = po$train(list(task))[[1L]]
out$head()
#>    passengers passengers_catch22_DN_HistogramMode_5 passengers_catch22_DN_HistogramMode_10 passengers_catch22_CO_f1ecac passengers_catch22_CO_FirstMin_ac
#>         <num>                                 <num>                                  <num>                        <num>                             <num>
#> 1:        112                              -1.03778                             -0.8218858                     28.49835                                 8
#> 2:        118                              -1.03778                             -0.8218858                     28.49835                                 8
#> 3:        132                              -1.03778                             -0.8218858                     28.49835                                 8
#> 4:        129                              -1.03778                             -0.8218858                     28.49835                                 8
#> 5:        121                              -1.03778                             -0.8218858                     28.49835                                 8
#> 6:        135                              -1.03778                             -0.8218858                     28.49835                                 8
#>    passengers_catch22_CO_HistogramAMI_even_2_5 passengers_catch22_CO_trev_1_num passengers_catch22_MD_hrv_classic_pnn40 passengers_catch22_SB_BinaryStats_mean_longstretch1
#>                                          <num>                            <num>                                   <num>                                               <num>
#> 1:                                   0.7175964                     -0.002839309                               0.8951049                                                  48
#> 2:                                   0.7175964                     -0.002839309                               0.8951049                                                  48
#> 3:                                   0.7175964                     -0.002839309                               0.8951049                                                  48
#> 4:                                   0.7175964                     -0.002839309                               0.8951049                                                  48
#> 5:                                   0.7175964                     -0.002839309                               0.8951049                                                  48
#> 6:                                   0.7175964                     -0.002839309                               0.8951049                                                  48
#>    passengers_catch22_SB_TransitionMatrix_3ac_sumdiagcov passengers_catch22_PD_PeriodicityWang_th0_01 passengers_catch22_CO_Embed2_Dist_tau_d_expfit_meandiff
#>                                                    <num>                                        <num>                                                   <num>
#> 1:                                             0.1666667                                           11                                               0.6712426
#> 2:                                             0.1666667                                           11                                               0.6712426
#> 3:                                             0.1666667                                           11                                               0.6712426
#> 4:                                             0.1666667                                           11                                               0.6712426
#> 5:                                             0.1666667                                           11                                               0.6712426
#> 6:                                             0.1666667                                           11                                               0.6712426
#>    passengers_catch22_IN_AutoMutualInfoStats_40_gaussian_fmmi passengers_catch22_FC_LocalSimple_mean1_tauresrat passengers_catch22_DN_OutlierInclude_p_001_mdrmd
#>                                                         <num>                                             <num>                                            <num>
#> 1:                                                          5                                        0.03846154                                        0.7847222
#> 2:                                                          5                                        0.03846154                                        0.7847222
#> 3:                                                          5                                        0.03846154                                        0.7847222
#> 4:                                                          5                                        0.03846154                                        0.7847222
#> 5:                                                          5                                        0.03846154                                        0.7847222
#> 6:                                                          5                                        0.03846154                                        0.7847222
#>    passengers_catch22_DN_OutlierInclude_n_001_mdrmd passengers_catch22_SP_Summaries_welch_rect_area_5_1 passengers_catch22_SB_BinaryStats_diff_longstretch0
#>                                               <num>                                               <num>                                               <num>
#> 1:                                       -0.6736111                                           0.9485463                                                   5
#> 2:                                       -0.6736111                                           0.9485463                                                   5
#> 3:                                       -0.6736111                                           0.9485463                                                   5
#> 4:                                       -0.6736111                                           0.9485463                                                   5
#> 5:                                       -0.6736111                                           0.9485463                                                   5
#> 6:                                       -0.6736111                                           0.9485463                                                   5
#>    passengers_catch22_SB_MotifThree_quantile_hh passengers_catch22_SC_FluctAnal_2_rsrangefit_50_1_logi_prop_r1 passengers_catch22_SC_FluctAnal_2_dfa_50_1_2_logi_prop_r1
#>                                           <num>                                                          <num>                                                     <num>
#> 1:                                      1.50827                                                      0.7692308                                                 0.2820513
#> 2:                                      1.50827                                                      0.7692308                                                 0.2820513
#> 3:                                      1.50827                                                      0.7692308                                                 0.2820513
#> 4:                                      1.50827                                                      0.7692308                                                 0.2820513
#> 5:                                      1.50827                                                      0.7692308                                                 0.2820513
#> 6:                                      1.50827                                                      0.7692308                                                 0.2820513
#>    passengers_catch22_SP_Summaries_welch_rect_centroid passengers_catch22_FC_LocalSimple_mean3_stderr
#>                                                  <num>                                          <num>
#> 1:                                          0.02454369                                      0.4030752
#> 2:                                          0.02454369                                      0.4030752
#> 3:                                          0.02454369                                      0.4030752
#> 4:                                          0.02454369                                      0.4030752
#> 5:                                          0.02454369                                      0.4030752
#> 6:                                          0.02454369                                      0.4030752