Creat Lags of Target Variable

...

Parameters

The parameters are the parameters inherited from mlr3pipelines::PipeOpTaskPreproc, as well as the following parameters:

lags :: integer()
The lags to create.

Super classes

mlr3pipelines::PipeOp -> mlr3pipelines::PipeOpTaskPreproc -> PipeOpFcstLag

Methods

Public methods

PipeOpFcstLag$new()

PipeOpFcstLag$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Initializes a new instance of this Class.

Usage

PipeOpFcstLag$new(id = "fcst.lags", param_vals = list())

Arguments

id

(character(1))
Identifier of resulting object, default "fcst.lags".

param_vals

(named list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction. Default list().

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpFcstLag$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

set.seed(1234L)